Squash warnings
svn path=/trunk/libcss/; revision=6149

+++ test/parse2-auto.c
@@ -188,9 +188,9 @@ void run_test(const uint8_t *data, size_t len, const char *exp, size_t explen)
 	dump_sheet(sheet, buf, &buflen);
 
 	if (2 * explen - buflen != explen || memcmp(buf, exp, explen) != 0) {
-		printf("Expected (%zu):\n%.*s\n", explen, explen, exp);
+		printf("Expected (%zu):\n%.*s\n", explen, (int) explen, exp);
 		printf("Result (%zu):\n%.*s\n", 2 * explen - buflen,
-			2 * explen - buflen, buf);
+			(int) (2 * explen - buflen), buf);
 		assert(0 && "Result doesn't match expected");
 	}
